requirements.txt
openai
python-dotenv
streamlit
tiktoken
requests
fpdf2
reportlab
